Added pause on headset unplug Fixed visibility of menu_list
