+++ app/src/main/res/menu/menu_list.xml
@@ -6,13 +6,11 @@
         android:id="@+id/action_sort"
         android:icon="@drawable/ic_sort_white_24dp"
         android:title="@string/title_menu_sort"
-        android:visible="false"
         app:showAsAction="ifRoom" />
     <item
         android:id="@+id/action_search"
         android:icon="@drawable/ic_search_white_24dp"
         android:title="@string/title_menu_search"
         app:actionViewClass="android.support.v7.widget.SearchView"
-        android:visible="false"
         app:showAsAction="ifRoom|collapseActionView" />
 </menu>
